making integer integers

+++ Simulation_engine/simulate_pop_from_reform.py
@@ -196,7 +196,7 @@ def compare(period: str, dictionnaire_simulations, compute_deciles=True):
         dic_res_brut = impots_par_reforme.to_dict()
         for simu in dic_res_brut:
             for cas_type in dic_res_brut[simu]:
-                dic_res_brut[simu][cas_type] = round(dic_res_brut[simu][cas_type])
+                dic_res_brut[simu][cas_type] = int(round(dic_res_brut[simu][cas_type]))
         resultat = {"total": total, "res_brut": dic_res_brut}
 
     return resultat
@@ -476,7 +476,7 @@ def foyertodictcastype(idfoy, data=None):
     )
 
     dicres = {
-        "revenu": round(revenu),
+        "revenu": int(round(revenu)),
         "nombre_declarants": int(nbpr),
         "nombre_personnes_a_charge": int(nbpac),
         "nombre_declarants_retraites": int(nbret),
